# services/api-service/internal/swaggerui/openapi.yaml
openapi: 3.0.3
info:
  title: ASOC Public API (api-service)
  description: |
    **api-service.** `POST /api/v1/scans` — единый вход: в теле задаётся **`scanner_id`** (`semgrep`, `gitleaks`, …), далее тот же сценарий (сканирование, ingest находок, группы, тикеты).
    `POST /api/v1/scans/semgrep`, `…/gitleaks`, `…/sca` (Trivy) и `…/dast` (OWASP ZAP) оставлены для совместимости (тело как при соответствующем сканере — без поля `scanner_id`).
    **`POST /api/v1/findings/ingest`** — приём **уже нормализованных** находок в формате ниже (CI, сторонний сканер); поле `owner_user_id` в теле **не используется**, владелец берётся только из JWT пользователя консоли при наличии. Поле **`console_product_id`** (опционально) привязывает прогон к записи **`GET /api/v1/console/products`** того же пользователя; без JWT отбрасывается. Отчёт и группы по одному проекту — **`GET /api/v1/report/vulnerabilities?console_product_id=…`** и **`GET /api/v1/groups?console_product_id=…`** (нужен JWT пользователя, не только API-ключ). С внешнего контура авторизуйте запрос тем же **`X-API-Key` / Bearer API-ключа** или JWT (см. `APP_AUTH_API_KEY`).
    `GET /api/v1/integrations` — публичный каталог сканеров (без секретов): метаданные, флаг доступности для консоли.
    Способ передачи находок и настройки среды — в `docs/ENVIRONMENT.md` и `docs/ARCHITECTURE.md`.
  version: "1.0"
servers:
  - url: https://atomic-asoc.ru
    description: Стенд (через публичный вход / asoc-web)
  - url: http://localhost:8080
...
paths:
  /api/v1/integrations:
    get:
      summary: Каталог интеграций (сканеров)
      operationId: listIntegrations
      responses:
        "200":
          description: Список инструментов
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntegrationsListResponse"
        "405":
          description: Метод не разрешён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/v1/admin/integrations:
    get:
      summary: Каталог интеграций для Atomic-admin
      operationId: adminListIntegrations
      responses:
        "200":
          description: Встроенные и дополнительные сканеры
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminIntegrationsCatalogDetail"
        "403":
          description: Нужен JWT роли admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "405":
          description: Метод не разрешён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
    put:
      summary: Замена «дополнительных» сканеров
      operationId: adminReplaceIntegrations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminIntegrationsReplaceBody"
      responses:
        "200":
          description: Применено
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
        "400":
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "403":
          description: Нужен JWT роли admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "405":
          description: Метод не разрешён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
...
  /api/v1/findings/ingest:
    post:
      summary: Ingest находок произвольного сканера (нормализованный JSON)
      description: >
        Отправка пакета в тот же контракт, что internal `processing-service`; маршрутизация через api-service (Kafka или HTTP по ENV).
        Пустой массив `findings` допустим — создаётся run без строк.
        Поле `channel` (manual/ci): если не указано — подставляется **ci** (ожидаемый сценарий из CI/CD).
      operationId: ingestFindingsPublic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessingIngestRequestPublic"
      responses:
        "202":
          description: Пакет принят (Kafka); при брокере — status/correlation_id, без run_id
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/IngestAccepted"
                  - $ref: "#/components/schemas/ProcessingSummary"
        "400":
          description: Неверное тело
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "401":
          description: Нет API-ключа / JWT для /api/*
        "502":
          description: Ошибка Kafka / processing-service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/v1/scans:
    post:
      summary: Полный сценарий сканирования (dispatch по scanner_id)
      operationId: scanUnified
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnifiedScanRequest"
      responses:
        "202":
          description: Паспорт сценария
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PassportResponse"
        "400":
          description: Неверное тело (в т.ч. неизвестный scanner_id или пустая цель)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "502":
          description: Ошибка downstream
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/v1/scans/semgrep:
    post:
      deprecated: true
      summary: Полный сценарий сканирования Semgrep (устарело — используйте POST /api/v1/scans и scanner_id)
      operationId: scanSemgrepLegacy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanRequest"
      responses:
        "202":
          description: Паспорт сценария
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PassportResponse"
        "400":
          description: Неверное тело
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "502":
          description: Ошибка downstream
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/v1/scans/sca:
    post:
      summary: Полный сценарий SCA (Trivy; без scanner_id в теле)
      operationId: scanScaCompat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanRequest"
      responses:
        "202":
          description: Паспорт сценария
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PassportResponse"
        "400":
          description: Неверное тело
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "502":
          description: Ошибка downstream
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/v1/scans/dast:
    post:
      summary: Полный сценарий DAST (OWASP ZAP baseline; без scanner_id в теле)
      operationId: scanDastCompat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanRequest"
      responses:
        "202":
          description: Паспорт сценария
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PassportResponse"
        "400":
          description: Неверное тело
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "502":
          description: Ошибка downstream
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/v1/scans/gitleaks:
    post:
      summary: Полный сценарий сканирования Gitleaks (без scanner_id в теле)
      operationId: scanGitleaksCompat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanRequest"
      responses:
        "202":
          description: Паспорт сценария
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PassportResponse"
        "400":
          description: Неверное тело
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "502":
          description: Ошибка downstream
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
components:
  schemas:
    IntegrationsListResponse:
      type: object
      properties:
        integrations:
          type: array
          items:
            $ref: "#/components/schemas/IntegrationCatalogItem"
    AdminIntegrationsCatalogDetail:
      type: object
      properties:
        overlay_persistent:
          type: boolean
          description: true если задан APP_INTEGRATIONS_OVERLAY_PATH (данные сохраняются между рестартами)
        entries:
          type: array
          items:
            $ref: "#/components/schemas/AdminCatalogEntry"
    AdminCatalogEntry:
      type: object
      properties:
        source:
          type: string
          enum: [builtin, additional]
        integration:
          $ref: "#/components/schemas/IntegrationCatalogItem"
    AdminIntegrationsReplaceBody:
      type: object
      required: [additional]
      properties:
        additional:
          type: array
          items:
            $ref: "#/components/schemas/IntegrationCatalogItem"
    IntegrationCatalogItem:
      type: object
      properties:
        id:
          type: string
        kind:
          type: string
          description: SAST, SCA, DAST и т.д.
        title:
          type: string
        summary:
          type: string
        phase:
          type: string
          enum: [ready, planned]
        enabled:
          type: boolean
        input_kind:
          type: string
          description: filesystem, lockfile, http, oci_image, http_target, manifest_only
        required_context:
          type: array
          items:
            type: string
          description: scm, base_url — что должно быть у продукта до запуска
        scanner_name:
          type: string
        api_scan_path:
          type: string
        console_path:
          type: string
        capabilities:
          type: array
          items:
            type: string
        note:
          type: string
        network_host:
          type: string
        network_hostname:
          type: string
        network_ip:
          type: string
        network_port:
          type: string
        scanner_invoke_url:
          type: string
          description: Полный URL POST (универсальный раннер /api/v1/run или свой адаптер).
        runner_command:
          type: string
          description: Шаблон shell для generic-scan-runner; плейсхолдеры {target_path}, {git_repository_url} и др.
        invoke_payload_template:
          type: string
          description: Пример или шаблон тела POST для вызова сканера.
        invoke_hint:
          type: string
    ErrorBody:
      type: object
      properties:
        error:
          type: string
    ProcessingFindingItem:
      type: object
      required:
        - asset_id
        - identifier
        - severity
        - component
        - version
        - cve
        - cwe
      properties:
        asset_id:
          type: string
        identifier:
          type: string
        severity:
          type: string
        component:
          type: string
        version:
          type: string
        cve:
          type: string
        cwe:
          type: string
        metadata:
          type: object
          additionalProperties: true
        raw_payload:
          type: object
          additionalProperties: true
    ProcessingIngestRequestPublic:
      type: object
      required:
        - scanner_name
        - findings
      properties:
        scanner_name:
          type: string
          description: Произвольное имя источника (отображается в processing run).
          example: my-custom-sast
        findings:
          type: array
          items:
            $ref: "#/components/schemas/ProcessingFindingItem"
        owner_user_id:
          type: integer
          format: int64
          deprecated: true
          description: Игнорируется; владелец задаётся только JWT консоли на api-service.
        channel:
          type: string
          enum: [manual, ci]
          description: >
            Канал прогона. Если не указано, для POST /api/v1/findings/ingest подставляется «ci» (типичный сценарий пайплайна).
            Ручные сканы через POST /api/v1/scans всегда фиксируются как «manual» на стороне сервера.
          example: ci
        console_product_id:
          type: integer
          format: int64
          description: >
            Опционально: id строки из GET /api/v1/console/products (продукт того же пользователя JWT).
            Прогон сохраняется с привязкой к проекту; GET /api/v1/groups и GET /api/v1/report/vulnerabilities с query-параметром
            console_product_id возвращают только данные прогонов этого продукта. С одним API-ключом без JWT поле отбрасывается.
    UnifiedScanRequest:
      allOf:
        - type: object
          required:
            - scanner_id
          properties:
            scanner_id:
              type: string
              example: semgrep
              description: Идентификатор исполнителя (`semgrep`, `gitleaks`, `trivy-sca`, `zap-dast`, …).
            options:
              type: object
              additionalProperties: true
              description: Необязательные параметры сканера (зарезервировано; Semgrep их пока не использует).
        - $ref: "#/components/schemas/ScanRequestBodyFields"
    ScanRequest:
      description: Совместимо с телом UnifiedScanRequest без поля scanner_id (legacy `/scans/semgrep`).
      allOf:
        - $ref: "#/components/schemas/ScanRequestBodyFields"
    ScanRequestBodyFields:
      type: object
      properties:
        scanner_name:
          type: string
          example: semgrep
        git_repository_url:
          type: string
          description: >
            Если задан, semgrep-service сначала выполняет shallow clone этого remote.
            Допустимы http(s) и форма git@host:path/repo.git; без приватных секретов в составе платформы
            поддерживаются в основном публичные репозитории.
        git_repository_ref:
          type: string
          example: main
          description: Ветка или тег для clone (--branch; при ошибке — clone и git checkout ref).
        target_path:
          type: string
          example: /app/demo/scan-targets/WebGoat/
          description: >
            Без git_repository_url — абсолютный каталог исходников внутри контейнера semgrep-service.
            С git_repository_url — относительный подкаталог внутри клона (пустая строка — корень репозитория).
        target_url:
          type: string
          example: https://example.com/app
          description: Публичный HTTP(S) URL для DAST (zap-dast); localhost запрещён.
        semgrep_config:
          type: string
          example: p/java
          description: Путь к YAML правилам или набор Semgrep (p/java)
    IngestAccepted:
      type: object
      properties:
        status:
          type: string
          example: accepted
        correlation_id:
          type: string
          format: uuid
        findings_queued:
          type: integer
    ProcessingSummary:
      type: object
      properties:
        run_id:
          type: integer
          format: int64
        findings_received:
          type: integer
        findings_processed:
          type: integer
        vulnerabilities_created:
          type: integer
        groups_updated:
          type: integer
    GroupItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
        group_key:
          type: string
        grouping_rule:
          type: string
        severity_max:
          type: string
        assets_count:
          type: integer
        status:
          type: string
    TicketItem:
      type: object
      properties:
        group_id:
          type: integer
          format: int64
        jira_issue_key:
          type: string
        jira_issue_url:
          type: string
        sync_status:
          type: string
        idempotency_key:
          type: string
    PassportResponse:
      type: object
      properties:
        scanner_name:
          type: string
        scan_target:
          type: string
        findings:
          type: array
          items: {}
        processing:
          $ref: "#/components/schemas/ProcessingSummary"
        groups:
          type: array
          items:
            $ref: "#/components/schemas/GroupItem"
        tickets:
          type: array
          items:
            $ref: "#/components/schemas/TicketItem"